Multiple Changes - Tests, Segfaults
Fixes for segfaults in lookup and catalog methods. Currently several
limitations on what can be inserted into the Trie - presently, just case
insensitive alphabetic characters. This will be changed in the future to
support all ASCII (and maybe UTF-8?) characters.

A Makefile for tests now available. Still tweaking the Makefiles to
support debug symbols and more options. Trying to keep it generic,
however.